requirements.txt
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2==2.9.9
djangorestframework-simplejwt==5.3.1
django-filter==24.2

